+++ src/main/java/alumnithon/skilllink/domain/userprofile/repository/CountryRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import alumnithon.skilllink.domain.userprofile.model.Country;
 @Repository
-public interface CountryRepository extends JpaRepository<Country, Long>{
+public interface CountryRepository extends JpaRepository<Country, Integer>{
 
-    Country getById(long countryId);
+    Country getById(Integer countryId);
 
 }
